Refactor and clean up e2e framework utils, this patch handles test/e2e/framework/kubelet_stats.go file
