+++ test/e2e/node/node_problem_detector.go
@@ -28,6 +28,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2ekubelet "k8s.io/kubernetes/test/e2e/framework/kubelet"
 	e2elog "k8s.io/kubernetes/test/e2e/framework/log"
 	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
 	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
@@ -266,7 +267,7 @@ func getCPUStat(f *framework.Framework, host string) (usage, uptime float64) {
 }
 
 func getNpdPodStat(f *framework.Framework, nodeName string) (cpuUsage, rss, workingSet float64) {
-	summary, err := framework.GetStatsSummary(f.ClientSet, nodeName)
+	summary, err := e2ekubelet.GetStatsSummary(f.ClientSet, nodeName)
 	framework.ExpectNoError(err)
 
 	hasNpdPod := false
